profiles: Add shared profile
The shared profile will store all global variables shared across each user accessing
it's local katalon project. On the other hand the default profile is private, meaning
it's not shared to the repository. The default profile can be used for custom global
variables which have meaning to only the user who creates them.

Signed-off-by: João Lourenço <jlourenco5691@gmail.com>